Refactor UI and ListView
Streamline the UI by removing unnecessary use of Label. The ListView is now responsible for displaying and handling the notes. It has methods for getting the selected item and/or index removing the need for variables tracking those.

All new notes now have a a default title to avoid no title being displayed if user doesn't update it.